divvunspell::speller
Trait Speller

pub trait Speller {
    // Required methods
    fn is_correct(self: Arc<Self>, word: &str) -> bool;
    fn is_correct_with_config(
        self: Arc<Self>,
        word: &str,
        config: &SpellerConfig,
    ) -> bool;
    fn suggest(self: Arc<Self>, word: &str) -> Vec<Suggestion>;
    fn suggest_with_config(
        self: Arc<Self>,
        word: &str,
        config: &SpellerConfig,
    ) -> Vec<Suggestion>;
}

can determine if string is a correct word or suggest corrections. Also with SpellerConfig.

Required Methods

fn is_correct(self: Arc<Self>, word: &str) -> bool

check if the word is correctly spelled

fn is_correct_with_config( self: Arc<Self>, word: &str, config: &SpellerConfig, ) -> bool

check if word is correctly spelled with config recasing etc.

fn suggest(self: Arc<Self>, word: &str) -> Vec<Suggestion>

suggest corrections to word

fn suggest_with_config( self: Arc<Self>, word: &str, config: &SpellerConfig, ) -> Vec<Suggestion>

suggest corrections with recasing and reweighting from config