Module divvunspell::tokenizer::word

Unicode Words of a string.

References

https://www.unicode.org/reports/tr29/#Word_Boundaries

Structs

WordBoundIndices

External iterator for word boundaries and byte offsets.

WordBounds

External iterator for a string’s word boundaries.

Words

An iterator over the substrings of a string which, after splitting the string on word boundaries, contain any characters with the Alphabetic property, or with General_Category=Number.